[ADM] Add type field to modelUpdated events and additional data.
Note: This changes the meaning of the 'node' parameter in some of the events.

The purpose is to make it easier to listen to all the events for a design,
but weed out ones you don't care about or find out more detail. Since Shane
is not currently using the events like childAdded we can probably remove all
those, especially with this patch.