[Widgets] Add htmlAttribute and forceAttribute property features
Changed how the "attribute map" works in builder.js. Instead of assuming
that every property corresponds to an HTML attribute, it looks for a
specific field in the property object "htmlAttribute". If this is present,
it contains the name of the attribute this property should be written to.

Also changed the behavior so that if this value is equal to the default
value for the property, we skip writing the attribute at all. However the
new "forceAttribute" field in the property object can override this
behavior if set to true, and then the attribute will be written.

This significantly cleans up our HTML output because lots of attributes
were being written that were completely bogus, and others were written when
they didn't need to be because they represented normal behavior.

Also replaced Slider template with a template function because the patch
broke it and this was the easiest way to fix it.

Also split ListView into two widgets: List and OrderedList, and ListItem
into ListItem and ListDivider. This removed the need for conditional logic,
and I thought they were natural widget-level distinctions for the user to see.

Left the Button widget untouched, which means that its probably rather broken
because its properties are being ignored. But the id and text properties work,
which covers 95% of Button use so I thought that could wait.