[General] Remove message about selecting append point within nested tags
Previously there was code here doing a hard-coded check for Radio Buttons or
Checkboxes and then appending to the child of the current DOM node instead
of the node itself. This was a nasty hack. The right solution would be to
provide a selector for a widget (or property) that lets the code select the
appropriate modification point beneath.

I happened to be temporarily removing the outer wrapper div from those two
widgets that led to this hack, so I just removed the hack rather than come
up with the real solution, which is no longer needed at the moment. I just
wanted to make sure someone sees this before they add something like that
back in. Whether it will work, I have no idea.