[Serialize] Use correct document creation method
By using the createHTMLDocument() instead of createDocument()
we now get the correct (default) HTML namespacing and parsing
results.

The also saves us from having to "fixup" the strings to be
valid HTML (such as the "<script />" and "<meta />" tags we
were getting before as a result of being XHTML namespaced).

Signed-off-by: Shane Bryan <shane.bryan@intel.com>