[Widgets] Generalize the hacks of Select, button and FlipToggleSwith
This patch introduced the concept of delegation and events into widgets
definition.
Delegation means that a widget can use another node as its delegate so
that when in design view we can drag the delgate node around and make it
selected. This feature is needed because jQM genertes enhencement tags for
a widget and sometimes these enhencement tags make the original tag not
selectable or draggable. By definning a delegate, the desgin view can use
another DOM node, usually generated by jQM to respond to the click and drag
events.
The value of "delegate" attribute of a widget can be either "next", "parent",
"grandparent" or a function which can return a delegate node.
The events attribute comes into being because that some special widgets need
to bind some special handlers for some events. For example, SelectMenu needs
to route the click event to Option if the user selects an option;
FlipToggleSwitch needs to select itself when a "change" event occurs.