[ADM] Remove unused events on individual nodes
The childAdded, childRemoved, and propertyChange events sent on each
ADM node have never actually been used. Recently enough information was
added to the modelUpdated event that these distinct causes can be
determined from it, so we lose no power by removing them, and it's more
efficient to listen only on the design root for events anyway.