[Widgets] Add ListButton widget and htmlSelector to control attributes
The ListButton widget is a new type of list item that is clickable and has
a target like a Button. To implement this, we needed more control over
where an HTML attribute gets applied to the template.

Now a widget property definition with an htmlAttribute set can add an
htmlSelector field, which gives a string selector to run against the
template to find the element(s) on which the attribute should be applied.