[General] Force window resize event to fix layout of panels
Commit 637905c4 exposed a bug in our assumptions about when
and how the window load/layout completes and gets resized.

Now that we're not waiting for an asynconous ajax call to
finish before the load completes, our resize handlers for
the palette and properties is not getting fired.

So I've added an explicit "trigger" of the "resize" event on
the window right after we complete the app initialization.

Signed-off-by: Shane Bryan <shane.bryan@intel.com>