[Design View] Fix PTSDK-235: Drop fails on some containers...
Drop fails on some containers if Design View is scrolled

Turns out that when the iframe for the Design View is scrolled,
the offset calculation, when used to determine intersections, will
still not be valid as it does not account for the windows scrollX
and scrollY properties.

Now scrollX and scrollY are subtracted from the offset in our
customized version of jQuery.ui.ddmanager.prepareOffsets().

Also fixed up a bunch of formatting and variable hoisting issues
in the code while I was mucking around.

Signed-off-by: Shane Bryan <shane.bryan@intel.com>