[Widgets] Fix PTSDK-181: Radio-Button can't be checked
Converted RadioButton tempate to a function that applies
the necessary "name" property to each input in the fieldset.

The first RadioButton added defines the name to be the same
as it's id (which is already auto generated).  Subsequent
siblings will inherit their name property from the first
RadioButton in the parent ControlGroup node.

Signed-off-by: Shane Bryan <shane.bryan@intel.com>