[General] Start of GUI Builder plugin model refactoring
The main "app" will itself now be a jQuery $.widget plugin
that takes a document, empties it, and repopulates it with
the structure it needs to present the application.

Every attempt will be made from here on out to maintain an
MVC (Model, View, Controller) architecture between the ADM
(the model), the views (layout, code, live, outline, etc.)
and the controller (the "app", aka GUI Builder).

All Views will themselves be jQuery UI $.widget plugins,
each bound to an element and provided a reference to the
model to track changes (and send requests) to it.

All these plugins will occupy the $.gb namespace in jQuery

Many more changes to come, as this initial refactoring
commit has substantially changed how we construct the app
context and apply the view plugins, but has only just the
beginnings of the "hard" work, connecting ADM to all the
views, etc...

Cheers, and welcome to the brave new world...

Signed-off-by: Shane Bryan <shane.bryan@intel.com>