[Widgets] [ADM] Generate Block elements automatically for Grids
Use rows and columns properties on Grid to control the size, rather than
dragging Blocks manually. This handles all the subtype generation so the
user doesn't have to worry about it.

In the process, added new "init" function to widget definition which is
called when the widget is created (except when loading design from a file),
and gives an opportunity to generate children to match the default
properties. (Here two Blocks to match the 1 row / 2 col default.)

Also added new "setPropertyHook" function to widget property definition
which gets called just before a property changes, so the widget can react
(for Grid, this means adding or removing rows or columns of Blocks).

Outstanding issues:
1) This will create a new problem for undo/redo support which will have
   to be addressed. These row/column changes wouldn't be captured in the
   undo stack with the current implementation (on another branch).
2) Making a large change in number of rows or columns takes a long time
   because each new or removed Block is causing a modelUpdated which
   currently makes the design canvas do a full refresh. This could be
   resolved by finally making design canvas do incremental updates, or
   worked around by suppressing events here in some fashion and perhaps
   sending a new "restructure" event for the Grid node when all is done.
3) Testing the Grid I noticed large memory leaks, which this at least
   exacerbates since it's now so easy to create so many widgets. Needs
   more investigation.