[Serialize] Moved more functions into the $.gb namespace
Now use $.gb.getDesignHeaders() and $.gb.getDefaultHeaders() to
generate or get the current document headers for code or layout
views.

Also used new selector for layoutView (':gb-layoutView') in
serializeADMNodeToDOM() when attempting to find the parent DOM
element in the layoutView of the current ADM node.  This is a
more portable method of finding elements in the layoutView document
that is dynamically generated, and prevents the need to maintain
a global reference to it that is kept in sync.

Finally, fixed a bug where we only tested if the renderer passed
into serializeADMSubtreeToDOM was "null".  Turns out it is also
possible for it to be "undefined", so the correct test is just for
falsy-ness: if (renderer)...

Signed-off-by: Shane Bryan <shane.bryan@intel.com>