[Outline] Render Block elements that contain children as row/col labels
Outline now uses showInPalette to determine whether to display a node,
rather than "selectable". But a new feature in a widget definition, an
"outlineLabel" function can return a non-empty string if it wishes for
the outline view to render a label in place of the widget.

We use this feature for Grid Blocks. The outlineLabel function is such
that they will only show up if they contain child elements, and the
label will be "Row M, Column N" to help identify this group of widgets
within the grid.

Added a new "label" class to CSS for outline elements that renders in
gray italic with a smaller font for these labels.