[Widgets] Use the undo transaction data feature to fix undo for Grids
When the Grid rows or columns property is lowered, this could cause
other widgets to be removed from the tree. There was no way to restore
those widgets in the event of an undo operation. Now the Grid's
property hook function stores an array of the removed blocks as
transaction data. If an undo happens, the hook is called again with
this same data supplied, and it uses it to reconstruct the Grid as
it was before the row/column removal.