[General] Rename 'template.*' to 'composer.*'
The name "template" no longer relates to what these files actually
do, and moreover, causes confusion now that we have started adding
page and project templates (pre-canned versions that can be copied
as starting point for new instances).

The name "composer" is more suggestive of what funtion these files
are intended to provide, that is, they add the necessary behaviors
and styling that allow for drag, drop, sort, select, highlighting,
etc... in the layout view composition area, or workspace.

Signed-off-by: Shane Bryan <shane.bryan@intel.com>