[General] Added URL option parser and initial debug option
Options passed on the query/search portion of the URL will be parsed
and validated against a whitelist of allowed option names.

Values are currently limited to strings whose pattern matches the
regex "/^[a-zA-z0-9]([_-]*)$/".  Any values not matching this regex
will cause the option to be skipped.  This is intended as to serve
as an initial and blunt filter against code injection.

Also added the initial "debug" option capability, including a basic
sub-option example for mouse coordinate tracking.

The result of the option parsing is a new object in the $.gb namespace,
$.gb.options.  If no options are passed, this object will be undefined.
Any options passed with become properties of $.gb.options, for example,
$.gb.options.debug.mousetrack=true, or $.gb.options.debug.loglevel=3.

More examples are given in the code comments.

Signed-off-by: Shane Bryan <shane.bryan@intel.com>