[Layout View] Completed porting of layout view...
Re-worked how palette and widget drag/drop/sorting is done,
dropping the jQuery-ui "droppable" plugin.  This simplified
the number of plugins trying to maintain state and control
the drag/drop behaviors.

For palette items, they continue to be ui.draggable widgets,
but now they are dynamically connected to the layout view
iframe ui.sortables, based on changes to the ADM.

And the sortable behavior is now fully selfcontained in the
composer.js code.

This change also got us the drop placeholder "ghosting" for
"free", so this commit also resolves PTSDK-262.  While I was
in that code, I also implemented PTSDK-261.

Signed-off-by: Shane Bryan <shane.bryan@intel.com>