[ADM] Add addChildRecursive() which retries add along ancestor chain
This is specifically useful for the paste() function. If you copy say
a Button, you have the button selected. If you immediately try to
paste, it didn't work because you can paste inside a button. But with
this function, we will try to paste it in the button's container, and
then its container, and so on until we succeed.