[Properties] Fix regressions introduced in b8715ef7
Commit b8715ef7 introduced two bugs:
- Page properties never shown
- Property change fails on all widgets except select/options

The first one is fixed by removing the check for event.node
being truthy, and by adding a check to allow "Page" widgets
to have thier properties shown.

The second is solved by adding back the assignment of local
element variable.  It was accidentaly removed in the offending
commit and resulted in an exception when trying to inkove a
method on a null/undefined object.

Signed-off-by: Shane Bryan <shane.bryan@intel.com>