[Project View] Rework project dialogs
Previously, none of the builtin jquery-ui buttons or events were
being utilized, causing duplication of functionality, and inconsitent
event handling

The dialog code not makes use of the default button capabilities
and updates the Title, text-input and button label on open.

We also now bind a "new-project-dialog" boolean data attribute to
the dialog that is set prior to calling it's open method.  This data
attribute is then used in the handler code to determine the text/action
to use/take.

Finally, we now also properly handle the <form> default submit action
so that if the user hits "enter" key while it has focus, we will call
the same close handler that gets called when they click the "Next" or
"Done" button.

Signed-off-by: Shane Bryan <shane.bryan@intel.com>