[General] Add hover/active states to various buttons
It's hard to tell things are clickable w/o hover states, and it's hard to
tell you've actually clicked if there is a delay, without active states.
So I've added color changes for these states to the add/dup page buttons
in the page picker, the settings/export buttons, and the tab selectors.

Because I couldn't find these things specified in the UX annotations
document, I marked them with /* GRG */ to try to indicate if you find
a better specification, feel free to overwrite these.