[Layout View] Fix sortable regression introduced in 7e851733
Unbinding of touchmove and mousedown caused unexpected, and as
yet undiagnosed, loss of mousedown events on pages not displayed
when pageinit occurs.

Further work is needed to isolate the cause and find a suitable
solution.

Until then, some widgets, like the color picker, are still able
to be manipulated in layout view when they should not be.

Signed-off-by: Shane Bryan <shane.bryan@intel.com>