[Layout View] Add header bars to right panel, remove incorrect borders
Rather than "Elements" header as in the mockups, I added "Widget Categories"
and "Widgets" but we can debate the wording later.

Also needed to fix the calculation of the heights of the variable panes in
the right panel. Figured out what the magic numbers meant and assigned them
to variables to make it clearer.