[General] Fixed cause of scrollHeight console error
Before the iframe document completes loading, we were querying
the iframes contentDocument height, which in turn tries to query
the contained body.scrollHeight.  Until the load completes, body
is null, thus the error.

I added a check in the resize handler to just return if the load
has not completed yet.

Signed-off-by: Shane Bryan <shane.bryan@intel.com>