[Tree View] Wait for selectionChanged event to update selection in outline
When the user clicks on a node in the outline, we need to call setSelected
in the ADM. But it will generate a selectionChanged event after which we
should update the selection highlight in the outline. We don't need to
also make that update at the time that we call ADM. For all we know, after
all, ADM will deny the request and not change the selection.