[Tree View] Fixed the problem when removing widgets from Grid, the label will not be removed in outline view
This problem is due to that we didn't delete the label, which is
created by _render function of outline view. To fix the problem,
I did the following thing:
1. Expecting _render to return newly created top level elements so
that tree view can delete it when removing a node.
2. Attach _origin_node to data of the returned node so that we can
find it when removing a node.
3. To maitain consistency, we'll attach _origin_data to folderNode
instead of the <a> node
4. Now that findDomNode returns differnt result, the callers of it
have to adapt to this change.