[ADM] Suppress events when autogenerating properties
This patch is triggered by a problem that when creating a new
page, there will be two duplicated page nodes in the outline view
of the Code tab. The root cause is that when generating the page
node of outline view, the page id property is needed. As it's
a newly created page, an id will be automatically generated, which
then triggers a propertyChanged event, which causes the outline
view to remove and then add the new node again. As there are two
outline views, the second outline view will call addNode twice
and an invalid removeNode once (at that time, there's no
corresponding node created for the second outline view). So two
nodes will be added to the second outline view.

Actually, no listener of ADM is interested in the generated
value as they are not using this property till now - if they
are using it, they must have called getProperty at least once,
which is impossible as auto generating only occurs the first time
getProperty is called for a given node's given property.