[Layout View] Fix PTSDK-586 - Button Group masked during sort
Turns out this bug also affected ANY adm-node element that was
also itself a sortable container (radio and cehckbox groups,
ordered and unordered lists, etc.).

Root cause was that we were never un-masking the sortable helper
element (and it's children) when it was itself a sortable.

Solution was to pass the ui.helper object to the applyMasking()
function and add it to the set of elements being un-masked.

At the same time, I adjusted the CSS for ui-unmasked class
elements to exclude those also classed as ui-sortable-helper.
This makes the helpers continue to have the desired appearance
as defined earlier in the CSS for .ui-sortable-helper elements
(specifically, to be 0.8 opaque and maintain their defaults for
the background).

Signed-off-by: Shane Bryan <shane.bryan@linux.intel.com>