[Serialize] Set properties with "raw" method to avoid block dups in grid
Resolves PTSDK-596.

The Grid element uses "property hook" functions to watch when the rows
or columns property changes, and then it responds by adding or removing
Block element children. This manages the blocks automatically so the
user doesn't have to worry about them. But when serializing, we need to
be able to set the row/column property we read from disk, without this
happening, because the Blocks were also serialized and will be added
manually as we parse the JSON tree.