[Widgets] Added 'value' property to textarea widget
Resolves PTSDK-219.

I'm not sure we should really have this at all, and less sure that we
should call it 'value', since for the text area the default text goes
between the open and close tag, rather than in a "value" attribute.
But it is the value as far as forms are concerned, so I'm just doing
this for consistency for now.