[General] Simplify the browser support dialog process
Only present one dialog (no need to mention the file://
command line args when the entire OS or Browser is not
supported).

Also switched the default action on the dialog to be to
NOT run the application.  We know the browser/OS is not
supported, so lets require them to cancel the default
action to proceed anyway

Finally, when they hit "OK", they will be redirected to
our project home page rather than left at a blank screen.

Signed-off-by: Shane Bryan <shane.bryan@linux.intel.com>