[FileIO] Allow imports from same file name
Previously, if the name of a project file selected to import is the same
as the one used during the last import, no onchange() event fired, so the
file would not import.

Now, the "file" input element is blanked out at the end of the "change"
event handler to allow importing files of the same name.

Signed-off-by: Shane Bryan <shane.bryan@intel.com>