[Widgets] Add new Text widget, removing the need for Label widget
The Text widget supports rendering as HTML text wrapped in a variety
of tags (span, h1 through h6, label, p, em, strong). Also modified the
palette categories to include the new Text widget aand remove Label.

Initially, I supported plain text as the default, with no tag at all,
and also supported a line break with the <br> tag. But that requires
wrapping the plain text in a temporary tag so the layout canvas has
something to work with, and the beautify script did not deal nicely
with plain text. Also, the 'id' field doesn't make sense for plain
text. So eventually I decided to just get in this 90% of the
functionality that works perfectly w/o the border cases. Further
refinements would be nice, but are less bang for the buck.