[Widgets] Use actual booleans for disabled property default values
Also, fix a minor inconsistency in the previous commit where the Slider
widget showed up disabled in the design canvas when disabled, while none
of the other widgets do.

At some point we should show all the widgets as disabled in the canvas
since that is more WYSIWYG. But we have to get around a problem where the
click events get swallowed. Left a FIXME note about that.