[Widgets] Fix bug in checked property on Radio Button
When a Radio Group is created, it includes three Radio Buttons by default;
the first one is checked. But its checked property was set incorrectly to
true, even though it's a string property. This caused the property view to
render it as "not checked" initially, even though it was checked and the
code was generated properly.

To catch this kind of bug faster, setProperty should check have some
type checking and warn about this.