[Widgets] [ADM] Mark getPropertySchema private and use only in BWidget
The intended pattern was that all information about properties would
be accessed through getter functions, so the details of the format of
the property schema can change without affecting anyone outside of
BWidget. A recent change to ADM hadn't followed this pattern.

Added a BWidget.propertyValidIn that uses the validIn and invalidIn
values, rather than accessing them directly. It now follows the
pattern of allowIn/denyIn as well, where either a single string or
an array of strings can be specified.

Also made a few minor style tweaks in widgets.js.