[Widgets] Rename "vertical button group" to "button group"
This solves the problem that you expect a horizontal button group
widget as well. Instead, the single button group can be set either
way.

We might want to later have vertical and horizontal palette icons
which create a Button Group that is preset to the correct
orientation, but this is an improvement for now.