[Widgets] Simplify definition of HTML attribute value mappings
Add a new "htmlValueMap" field for widget property definitions, and a
BWidget.getPropertyValueMap() to access it. This can be an object mapping
property values to HTML attribute values for serialization, or a function
that does this mapping.

Add a new BWidget.getPropertySerializableValue() which takes a given value
and performs the mapping above.

Finally, added an ADMNode.getPropertySerializationObject() which returns
the HTML attribute name, serializable value, and HTML selector for a
property, if it is appropriate to serialize. This eliminates the need for
getPropertyDomAttribute in serialize.js.