[Layout View] Use a new algorithm for sorting
Previous algorithm only supports one zone containers or Headers,
and it works hard to determine the position for insertion. The new
algorithm first tries to insert into the zone of the nearest
sibling using enhanced insertChildAfter/Before; if it fails, it uses
addChild to automatically find a zone to insert into.