[Layout View] Fix bug: Can't drag a ThumbnailSplitList
Actually this bug is caused by a bug in jQuery UI sortable's
_removeCurrentsFromItems. But if we make sure we don't have
items in the currentItem, this bug will not be triggered.
So I did the following to make this happen:
* Make sure a sortable doesnt connect with the currentItem
and its child sortables and  update the items correctly after
changing connectWith,
* Make sure the items selector doesn't select unwanted items -
The item selector of Navbar actaully makes Content select
elements in a ThumbnailSplitList as items.
This patch also fixes the problem that dragging a container
may cause DOM hierachy error. This is because if the children in
the currentItem is connected with the sortable, it's possible
that the placeholder be put into the children of the currentItem
and when dropping, jQuery UI will try to insert currentItem
after/before placeholder, which will cause a DOM hierachy error.